server/requirements.txt
flask-restplus
flask-sqlalchemy
flask-swagger-ui
Flask-RESTful
Flask-Cors
